Add "BruteForce"

+++ src/WorkingWithFiles.java
@@ -7,21 +7,31 @@ public class WorkingWithFiles {
 
     public List<String> readingFromFile(String url){
         Path path = Path.of(url);
-        List<String> list;
-        try {
-            list = Files.readAllLines(path);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+        List<String> list = null;
+        if (Files.exists(path)) {
+            try {
+                list = Files.readAllLines(path);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+            return list;
+        }else {
+            System.out.println("Неверный путь файла");
+            return list;
         }
-        return list;
+
     }
 
-    public void WritingToFile(List<String> list,String url){
+    public void WritingToFile(List<String> list,String url) {
         Path path = Path.of(url);
-        try {
-            Files.write(path,list);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+        if (Files.exists(path)) {
+            try {
+                Files.write(path, list);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }else {
+            System.out.println("Неверный путь файла");
         }
     }
